+++ db/migrate/20160421101817_create_items.rb
@@ -3,10 +3,10 @@ class CreateItems < ActiveRecord::Migration
     create_table :items do |t|
       t.references :project, index: true
       t.string :item_title
-      t.decimal :budget
-      t.decimal :amount_spent, default: 0
-      t.decimal :balance_left
-      t.decimal :percentage_spent, default: 0
+      t.float :budget
+      t.float :amount_spent, default: 0
+      t.float :balance_left
+      t.float :percentage_spent, default: 0
       t.text :comments
 
       t.timestamps null: false
